+++ ray-operator/controllers/ray/rayservice_controller_test.go
@@ -18,7 +18,6 @@ package ray
 import (
 	"context"
 	"fmt"
-	"os"
 	"time"
 
 	"github.com/ray-project/kuberay/ray-operator/controllers/ray/common"
@@ -48,41 +47,41 @@ var _ = Context("Inside the default namespace", func() {
 	var numReplicas int32 = 1
 	var numCpus float64 = 0.1
 
-	testServeAppName := "app1"
-	testServeConfigV2 := fmt.Sprintf(`
-applications:
-  - name: %s
-    import_path: fruit.deployment_graph
-    runtime_env:
-      working_dir: "https://github.com/ray-project/test_dag/archive/41d09119cbdf8450599f993f51318e9e27c59098.zip"
-    deployments:
-      - name: MangoStand
-        num_replicas: 1
-        user_config:
-          price: 3
-        ray_actor_options:
-          num_cpus: 0.1
-      - name: OrangeStand
-        num_replicas: 1
-        user_config:
-          price: 2
-        ray_actor_options:
-          num_cpus: 0.1
-      - name: PearStand
-        num_replicas: 1
-        user_config:
-          price: 1
-        ray_actor_options:
-          num_cpus: 0.1
-      - name: FruitMarket
-        num_replicas: 1
-        ray_actor_options:
-          num_cpus: 0.1
-      - name: DAGDriver
-        num_replicas: 1
-        route_prefix: "/"
-        ray_actor_options:
-          num_cpus: 0.1`, testServeAppName)
+	// testServeAppName := "app1"
+	// testServeConfigV2 := fmt.Sprintf(`
+	// applications:
+	//   - name: %s
+	//     import_path: fruit.deployment_graph
+	//     runtime_env:
+	//       working_dir: "https://github.com/ray-project/test_dag/archive/41d09119cbdf8450599f993f51318e9e27c59098.zip"
+	//     deployments:
+	//       - name: MangoStand
+	//         num_replicas: 1
+	//         user_config:
+	//           price: 3
+	//         ray_actor_options:
+	//           num_cpus: 0.1
+	//       - name: OrangeStand
+	//         num_replicas: 1
+	//         user_config:
+	//           price: 2
+	//         ray_actor_options:
+	//           num_cpus: 0.1
+	//       - name: PearStand
+	//         num_replicas: 1
+	//         user_config:
+	//           price: 1
+	//         ray_actor_options:
+	//           num_cpus: 0.1
+	//       - name: FruitMarket
+	//         num_replicas: 1
+	//         ray_actor_options:
+	//           num_cpus: 0.1
+	//       - name: DAGDriver
+	//         num_replicas: 1
+	//         route_prefix: "/"
+	//         ray_actor_options:
+	//           num_cpus: 0.1`, testServeAppName)
 
 	myRayService := &rayv1.RayService{
 		ObjectMeta: metav1.ObjectMeta{
@@ -323,344 +322,344 @@ applications:
 			}
 		})
 
-		It("Dashboard should be healthy", func() {
-			Eventually(
-				checkServiceHealth(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(BeTrue(), "My myRayService status = %v", myRayService.Status)
-		})
+		// It("Dashboard should be healthy", func() {
+		// 	Eventually(
+		// 		checkServiceHealth(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(BeTrue(), "My myRayService status = %v", myRayService.Status)
+		// })
 
-		It("should create a new head service resource", func() {
-			svc := &corev1.Service{}
-			headSvcName, err := utils.GenerateHeadServiceName(utils.RayServiceCRD, myRayService.Spec.RayClusterSpec, myRayService.Name)
-			Expect(err).To(BeNil(), "failed to generate head service name")
-			Eventually(
-				getResourceFunc(ctx, client.ObjectKey{Name: headSvcName, Namespace: "default"}, svc),
-				time.Second*15, time.Millisecond*500).Should(BeNil(), "My head service = %v", svc)
-			Expect(svc.Spec.Selector[common.RayIDLabelKey]).Should(Equal(utils.GenerateIdentifier(myRayCluster.Name, rayv1.HeadNode)))
-		})
+		// It("should create a new head service resource", func() {
+		// 	svc := &corev1.Service{}
+		// 	headSvcName, err := utils.GenerateHeadServiceName(utils.RayServiceCRD, myRayService.Spec.RayClusterSpec, myRayService.Name)
+		// 	Expect(err).To(BeNil(), "failed to generate head service name")
+		// 	Eventually(
+		// 		getResourceFunc(ctx, client.ObjectKey{Name: headSvcName, Namespace: "default"}, svc),
+		// 		time.Second*15, time.Millisecond*500).Should(BeNil(), "My head service = %v", svc)
+		// 	Expect(svc.Spec.Selector[common.RayIDLabelKey]).Should(Equal(utils.GenerateIdentifier(myRayCluster.Name, rayv1.HeadNode)))
+		// })
 
-		It("should create a new serve service resource", func() {
-			svc := &corev1.Service{}
-			Eventually(
-				getResourceFunc(ctx, client.ObjectKey{Name: utils.GenerateServeServiceName(myRayService.Name), Namespace: "default"}, svc),
-				time.Second*15, time.Millisecond*500).Should(BeNil(), "My serve service = %v", svc)
-			Expect(svc.Spec.Selector[common.RayClusterLabelKey]).Should(Equal(myRayCluster.Name))
-		})
+		// It("should create a new serve service resource", func() {
+		// 	svc := &corev1.Service{}
+		// 	Eventually(
+		// 		getResourceFunc(ctx, client.ObjectKey{Name: utils.GenerateServeServiceName(myRayService.Name), Namespace: "default"}, svc),
+		// 		time.Second*15, time.Millisecond*500).Should(BeNil(), "My serve service = %v", svc)
+		// 	Expect(svc.Spec.Selector[common.RayClusterLabelKey]).Should(Equal(myRayCluster.Name))
+		// })
 
-		It("should update a rayservice object and switch to new Ray Cluster", func() {
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
+		// It("should update a rayservice object and switch to new Ray Cluster", func() {
+		// 	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
 
-				myRayService.Spec.RayClusterSpec.RayVersion = "2.100.0"
-				return k8sClient.Update(ctx, myRayService)
-			})
+		// 		myRayService.Spec.RayClusterSpec.RayVersion = "2.100.0"
+		// 		return k8sClient.Update(ctx, myRayService)
+		// 	})
 
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
 
-			Eventually(
-				getPreparingRayClusterNameFunc(ctx, myRayService),
-				time.Second*15, time.Millisecond*500).Should(Not(BeEmpty()), "Pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
-			pendingRayClusterName := myRayService.Status.PendingServiceStatus.RayClusterName
+		// 	Eventually(
+		// 		getPreparingRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*15, time.Millisecond*500).Should(Not(BeEmpty()), "Pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
+		// 	pendingRayClusterName := myRayService.Status.PendingServiceStatus.RayClusterName
 
-			// Update the status of the head Pod to Running.
-			updateHeadPodToRunningAndReady(ctx, pendingRayClusterName)
+		// 	// Update the status of the head Pod to Running.
+		// 	updateHeadPodToRunningAndReady(ctx, pendingRayClusterName)
 
-			// Confirm switch to a new Ray Cluster.
-			Eventually(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*15, time.Millisecond*500).Should(Equal(pendingRayClusterName), "My new RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// 	// Confirm switch to a new Ray Cluster.
+		// 	Eventually(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*15, time.Millisecond*500).Should(Equal(pendingRayClusterName), "My new RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
 
-			Eventually(
-				getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Status.ActiveServiceStatus.RayClusterName, Namespace: "default"}, myRayCluster),
-				time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayCluster  = %v", myRayCluster.Name)
-		})
+		// 	Eventually(
+		// 		getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Status.ActiveServiceStatus.RayClusterName, Namespace: "default"}, myRayCluster),
+		// 		time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayCluster  = %v", myRayCluster.Name)
+		// })
 
-		It("Disable zero-downtime upgrade", func() {
-			// Disable zero-downtime upgrade.
-			os.Setenv("ENABLE_ZERO_DOWNTIME", "false")
+		// It("Disable zero-downtime upgrade", func() {
+		// 	// Disable zero-downtime upgrade.
+		// 	os.Setenv("ENABLE_ZERO_DOWNTIME", "false")
 
-			// Try to trigger a zero-downtime upgrade.
-			oldRayVersion := myRayService.Spec.RayClusterSpec.RayVersion
-			newRayVersion := "2.198.0"
-			Expect(oldRayVersion).ShouldNot(Equal(newRayVersion))
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
-				myRayService.Spec.RayClusterSpec.RayVersion = newRayVersion
-				return k8sClient.Update(ctx, myRayService)
-			})
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
+		// 	// Try to trigger a zero-downtime upgrade.
+		// 	oldRayVersion := myRayService.Spec.RayClusterSpec.RayVersion
+		// 	newRayVersion := "2.198.0"
+		// 	Expect(oldRayVersion).ShouldNot(Equal(newRayVersion))
+		// 	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
+		// 		myRayService.Spec.RayClusterSpec.RayVersion = newRayVersion
+		// 		return k8sClient.Update(ctx, myRayService)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
 
-			// Because the zero-downtime upgrade is disabled, the RayService controller will not prepare a new RayCluster.
-			Consistently(
-				getPreparingRayClusterNameFunc(ctx, myRayService),
-				time.Second*5, time.Millisecond*500).Should(BeEmpty(), "Pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
+		// 	// Because the zero-downtime upgrade is disabled, the RayService controller will not prepare a new RayCluster.
+		// 	Consistently(
+		// 		getPreparingRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*5, time.Millisecond*500).Should(BeEmpty(), "Pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
 
-			// Set the RayVersion back to the old value to avoid triggering the zero-downtime upgrade.
-			err = retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
-				myRayService.Spec.RayClusterSpec.RayVersion = oldRayVersion
-				return k8sClient.Update(ctx, myRayService)
-			})
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
+		// 	// Set the RayVersion back to the old value to avoid triggering the zero-downtime upgrade.
+		// 	err = retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
+		// 		myRayService.Spec.RayClusterSpec.RayVersion = oldRayVersion
+		// 		return k8sClient.Update(ctx, myRayService)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
 
-			// Enable zero-downtime upgrade again.
-			os.Unsetenv("ENABLE_ZERO_DOWNTIME")
+		// 	// Enable zero-downtime upgrade again.
+		// 	os.Unsetenv("ENABLE_ZERO_DOWNTIME")
 
-			// Zero-downtime upgrade should not be triggered.
-			Consistently(
-				getPreparingRayClusterNameFunc(ctx, myRayService),
-				time.Second*5, time.Millisecond*500).Should(BeEmpty(), "Pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
-		})
+		// 	// Zero-downtime upgrade should not be triggered.
+		// 	Consistently(
+		// 		getPreparingRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*5, time.Millisecond*500).Should(BeEmpty(), "Pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
+		// })
 
-		It("Autoscaler updates the active RayCluster and should not switch to a new RayCluster", func() {
-			// Simulate autoscaler by updating the active RayCluster directly. Note that the autoscaler
-			// will not update the RayService directly.
-			initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: initialClusterName, Namespace: "default"}, myRayCluster),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "Active RayCluster = %v", myRayCluster.Name)
-				podToDelete := workerPods.Items[0]
-				*myRayCluster.Spec.WorkerGroupSpecs[0].Replicas++
-				myRayCluster.Spec.WorkerGroupSpecs[0].ScaleStrategy.WorkersToDelete = []string{podToDelete.Name}
-				return k8sClient.Update(ctx, myRayCluster)
-			})
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayCluster")
+		// It("Autoscaler updates the active RayCluster and should not switch to a new RayCluster", func() {
+		// 	// Simulate autoscaler by updating the active RayCluster directly. Note that the autoscaler
+		// 	// will not update the RayService directly.
+		// 	initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
+		// 	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: initialClusterName, Namespace: "default"}, myRayCluster),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "Active RayCluster = %v", myRayCluster.Name)
+		// 		podToDelete := workerPods.Items[0]
+		// 		*myRayCluster.Spec.WorkerGroupSpecs[0].Replicas++
+		// 		myRayCluster.Spec.WorkerGroupSpecs[0].ScaleStrategy.WorkersToDelete = []string{podToDelete.Name}
+		// 		return k8sClient.Update(ctx, myRayCluster)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayCluster")
 
-			// Confirm not switch to a new RayCluster
-			Consistently(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*5, time.Millisecond*500).Should(Equal(initialClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
-			Eventually(
-				getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Status.ActiveServiceStatus.RayClusterName, Namespace: "default"}, myRayCluster),
-				time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayCluster  = %v", myRayCluster.Name)
-		})
+		// 	// Confirm not switch to a new RayCluster
+		// 	Consistently(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*5, time.Millisecond*500).Should(Equal(initialClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// 	Eventually(
+		// 		getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Status.ActiveServiceStatus.RayClusterName, Namespace: "default"}, myRayCluster),
+		// 		time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayCluster  = %v", myRayCluster.Name)
+		// })
 
-		It("Autoscaler updates the pending RayCluster and should not switch to a new RayCluster", func() {
-			// Simulate autoscaler by updating the pending RayCluster directly. Note that the autoscaler
-			// will not update the RayService directly.
+		// It("Autoscaler updates the pending RayCluster and should not switch to a new RayCluster", func() {
+		// 	// Simulate autoscaler by updating the pending RayCluster directly. Note that the autoscaler
+		// 	// will not update the RayService directly.
 
-			// Trigger a new RayCluster preparation by updating the RayVersion.
-			oldRayVersion := myRayService.Spec.RayClusterSpec.RayVersion
-			newRayVersion := "2.200.0"
-			Expect(oldRayVersion).ShouldNot(Equal(newRayVersion))
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
-				myRayService.Spec.RayClusterSpec.RayVersion = newRayVersion
-				return k8sClient.Update(ctx, myRayService)
-			})
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
-			Eventually(
-				getPreparingRayClusterNameFunc(ctx, myRayService),
-				time.Second*60, time.Millisecond*500).Should(Not(BeEmpty()), "New pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
-			initialPendingClusterName, _ := getPreparingRayClusterNameFunc(ctx, myRayService)()
+		// 	// Trigger a new RayCluster preparation by updating the RayVersion.
+		// 	oldRayVersion := myRayService.Spec.RayClusterSpec.RayVersion
+		// 	newRayVersion := "2.200.0"
+		// 	Expect(oldRayVersion).ShouldNot(Equal(newRayVersion))
+		// 	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
+		// 		myRayService.Spec.RayClusterSpec.RayVersion = newRayVersion
+		// 		return k8sClient.Update(ctx, myRayService)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
+		// 	Eventually(
+		// 		getPreparingRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*60, time.Millisecond*500).Should(Not(BeEmpty()), "New pending RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
+		// 	initialPendingClusterName, _ := getPreparingRayClusterNameFunc(ctx, myRayService)()
 
-			// Simulate that the pending RayCluster is updated by the autoscaler.
-			err = retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: initialPendingClusterName, Namespace: "default"}, myRayCluster),
-					time.Second*15, time.Millisecond*500).Should(BeNil(), "Pending RayCluster = %v", myRayCluster.Name)
-				podToDelete := workerPods.Items[0]
-				*myRayCluster.Spec.WorkerGroupSpecs[0].Replicas++
-				myRayCluster.Spec.WorkerGroupSpecs[0].ScaleStrategy.WorkersToDelete = []string{podToDelete.Name}
-				return k8sClient.Update(ctx, myRayCluster)
-			})
-			Expect(err).NotTo(HaveOccurred(), "Failed to update the pending RayCluster.")
+		// 	// Simulate that the pending RayCluster is updated by the autoscaler.
+		// 	err = retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: initialPendingClusterName, Namespace: "default"}, myRayCluster),
+		// 			time.Second*15, time.Millisecond*500).Should(BeNil(), "Pending RayCluster = %v", myRayCluster.Name)
+		// 		podToDelete := workerPods.Items[0]
+		// 		*myRayCluster.Spec.WorkerGroupSpecs[0].Replicas++
+		// 		myRayCluster.Spec.WorkerGroupSpecs[0].ScaleStrategy.WorkersToDelete = []string{podToDelete.Name}
+		// 		return k8sClient.Update(ctx, myRayCluster)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "Failed to update the pending RayCluster.")
 
-			// Confirm not switch to a new RayCluster when the pending RayCluster triggers autoscaler.
-			Consistently(
-				getPreparingRayClusterNameFunc(ctx, myRayService),
-				time.Second*5, time.Millisecond*500).Should(Equal(initialPendingClusterName), "Pending RayCluster name = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
+		// 	// Confirm not switch to a new RayCluster when the pending RayCluster triggers autoscaler.
+		// 	Consistently(
+		// 		getPreparingRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*5, time.Millisecond*500).Should(Equal(initialPendingClusterName), "Pending RayCluster name = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
 
-			// The pending RayCluster will become the active RayCluster after:
-			// (1) The pending RayCluster's head Pod becomes Running and Ready
-			// (2) The pending RayCluster's Serve Deployments are HEALTHY.
-			updateHeadPodToRunningAndReady(ctx, initialPendingClusterName)
-			fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
-			Eventually(
-				getPreparingRayClusterNameFunc(ctx, myRayService),
-				time.Second*15, time.Millisecond*500).Should(BeEmpty(), "Pending RayCluster name = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
-			Eventually(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*15, time.Millisecond*500).Should(Equal(initialPendingClusterName), "New active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
-		})
+		// 	// The pending RayCluster will become the active RayCluster after:
+		// 	// (1) The pending RayCluster's head Pod becomes Running and Ready
+		// 	// (2) The pending RayCluster's Serve Deployments are HEALTHY.
+		// 	updateHeadPodToRunningAndReady(ctx, initialPendingClusterName)
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
+		// 	Eventually(
+		// 		getPreparingRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*15, time.Millisecond*500).Should(BeEmpty(), "Pending RayCluster name = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
+		// 	Eventually(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*15, time.Millisecond*500).Should(Equal(initialPendingClusterName), "New active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// })
 
-		It("Status should be updated if the differences are not only LastUpdateTime and HealthLastUpdateTime fields.", func() {
-			// Make sure (1) Dashboard client is healthy (2) All the three Ray Serve deployments in the active RayCluster are HEALTHY.
-			initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
-			Eventually(
-				checkServiceHealth(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
+		// It("Status should be updated if the differences are not only LastUpdateTime and HealthLastUpdateTime fields.", func() {
+		// 	// Make sure (1) Dashboard client is healthy (2) All the three Ray Serve deployments in the active RayCluster are HEALTHY.
+		// 	initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
+		// 	Eventually(
+		// 		checkServiceHealth(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
 
-			// Change serve status to be unhealthy
-			fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.UNHEALTHY, rayv1.ApplicationStatusEnum.UNHEALTHY))
+		// 	// Change serve status to be unhealthy
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.UNHEALTHY, rayv1.ApplicationStatusEnum.UNHEALTHY))
 
-			// Confirm not switch to a new RayCluster.
-			Consistently(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(Equal(initialClusterName), "Active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// 	// Confirm not switch to a new RayCluster.
+		// 	Consistently(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(Equal(initialClusterName), "Active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
 
-			// Check if all the deployment statuses are UNHEALTHY.
-			checkAllDeploymentStatusesUnhealthy := func(ctx context.Context, rayService *rayv1.RayService) bool {
-				if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Name, Namespace: rayService.Namespace}, rayService); err != nil {
-					return false
-				}
-				for _, appStatus := range rayService.Status.ActiveServiceStatus.Applications {
-					for _, deploymentStatus := range appStatus.Deployments {
-						if deploymentStatus.Status != rayv1.DeploymentStatusEnum.UNHEALTHY {
-							return false
-						}
-					}
-				}
-				return true
-			}
+		// 	// Check if all the deployment statuses are UNHEALTHY.
+		// 	checkAllDeploymentStatusesUnhealthy := func(ctx context.Context, rayService *rayv1.RayService) bool {
+		// 		if err := k8sClient.Get(ctx, client.ObjectKey{Name: rayService.Name, Namespace: rayService.Namespace}, rayService); err != nil {
+		// 			return false
+		// 		}
+		// 		for _, appStatus := range rayService.Status.ActiveServiceStatus.Applications {
+		// 			for _, deploymentStatus := range appStatus.Deployments {
+		// 				if deploymentStatus.Status != rayv1.DeploymentStatusEnum.UNHEALTHY {
+		// 					return false
+		// 				}
+		// 			}
+		// 		}
+		// 		return true
+		// 	}
 
-			// The status update not only includes the LastUpdateTime and HealthLastUpdateTime fields, but also the ServeStatuses[i].Status field.
-			// Hence, all the ServeStatuses[i].Status should be updated to UNHEALTHY.
-			//
-			// Note: LastUpdateTime/HealthLastUpdateTime will be overwritten via metav1.Now() in rayservice_controller.go.
-			// Hence, we cannot use `newTime` to check whether the status is updated or not.
-			Eventually(
-				checkAllDeploymentStatusesUnhealthy(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
+		// 	// The status update not only includes the LastUpdateTime and HealthLastUpdateTime fields, but also the ServeStatuses[i].Status field.
+		// 	// Hence, all the ServeStatuses[i].Status should be updated to UNHEALTHY.
+		// 	//
+		// 	// Note: LastUpdateTime/HealthLastUpdateTime will be overwritten via metav1.Now() in rayservice_controller.go.
+		// 	// Hence, we cannot use `newTime` to check whether the status is updated or not.
+		// 	Eventually(
+		// 		checkAllDeploymentStatusesUnhealthy(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
 
-			fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
 
-			// Confirm not switch to a new RayCluster.
-			Consistently(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(Equal(initialClusterName), "Active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// 	// Confirm not switch to a new RayCluster.
+		// 	Consistently(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(Equal(initialClusterName), "Active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
 
-			// The status update not only includes the LastUpdateTime and HealthLastUpdateTime fields, but also the ServeStatuses[i].Status field.
-			// Hence, the status should be updated.
-			Eventually(
-				checkServiceHealth(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
-		})
+		// 	// The status update not only includes the LastUpdateTime and HealthLastUpdateTime fields, but also the ServeStatuses[i].Status field.
+		// 	// Hence, the status should be updated.
+		// 	Eventually(
+		// 		checkServiceHealth(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
+		// })
 
-		It("Status should not be updated if the only differences are the LastUpdateTime and HealthLastUpdateTime fields.", func() {
-			// Make sure (1) Dashboard client is healthy (2) All the three Ray Serve deployments in the active RayCluster are HEALTHY.
-			initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
-			Eventually(
-				checkServiceHealth(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
+		// It("Status should not be updated if the only differences are the LastUpdateTime and HealthLastUpdateTime fields.", func() {
+		// 	// Make sure (1) Dashboard client is healthy (2) All the three Ray Serve deployments in the active RayCluster are HEALTHY.
+		// 	initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
+		// 	Eventually(
+		// 		checkServiceHealth(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
 
-			// Only update the LastUpdateTime and HealthLastUpdateTime fields in the active RayCluster.
-			oldTime := myRayService.Status.ActiveServiceStatus.Applications[common.DefaultServeAppName].HealthLastUpdateTime.DeepCopy()
-			fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
+		// 	// Only update the LastUpdateTime and HealthLastUpdateTime fields in the active RayCluster.
+		// 	oldTime := myRayService.Status.ActiveServiceStatus.Applications[common.DefaultServeAppName].HealthLastUpdateTime.DeepCopy()
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
 
-			// Confirm not switch to a new RayCluster
-			Consistently(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(Equal(initialClusterName), "Active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// 	// Confirm not switch to a new RayCluster
+		// 	Consistently(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(Equal(initialClusterName), "Active RayCluster name = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
 
-			// The status is still the same as before.
-			Eventually(
-				checkServiceHealth(ctx, myRayService),
-				time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
+		// 	// The status is still the same as before.
+		// 	Eventually(
+		// 		checkServiceHealth(ctx, myRayService),
+		// 		time.Second*3, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
 
-			// Status should not be updated if the only differences are the LastUpdateTime and HealthLastUpdateTime fields.
-			// Unlike the test "Status should be updated if the differences are not only LastUpdateTime and HealthLastUpdateTime fields.",
-			// the status update will not be triggered, so we can check whether the LastUpdateTime/HealthLastUpdateTime fields are updated or not by `oldTime`.
-			Expect(myRayService.Status.ActiveServiceStatus.Applications[common.DefaultServeAppName].HealthLastUpdateTime).Should(Equal(oldTime), "myRayService status = %v", myRayService.Status)
-		})
+		// 	// Status should not be updated if the only differences are the LastUpdateTime and HealthLastUpdateTime fields.
+		// 	// Unlike the test "Status should be updated if the differences are not only LastUpdateTime and HealthLastUpdateTime fields.",
+		// 	// the status update will not be triggered, so we can check whether the LastUpdateTime/HealthLastUpdateTime fields are updated or not by `oldTime`.
+		// 	Expect(myRayService.Status.ActiveServiceStatus.Applications[common.DefaultServeAppName].HealthLastUpdateTime).Should(Equal(oldTime), "myRayService status = %v", myRayService.Status)
+		// })
 
-		It("Update workerGroup.replicas in RayService and should not switch to new Ray Cluster", func() {
-			// Certain field updates should not trigger new RayCluster preparation, such as updates
-			// to `Replicas` and `WorkersToDelete` triggered by the autoscaler during scaling up/down.
-			// See the function `generateRayClusterJsonHash` for more details.
-			initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
-				*myRayService.Spec.RayClusterSpec.WorkerGroupSpecs[0].Replicas++
-				return k8sClient.Update(ctx, myRayService)
-			})
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
+		// It("Update workerGroup.replicas in RayService and should not switch to new Ray Cluster", func() {
+		// 	// Certain field updates should not trigger new RayCluster preparation, such as updates
+		// 	// to `Replicas` and `WorkersToDelete` triggered by the autoscaler during scaling up/down.
+		// 	// See the function `generateRayClusterJsonHash` for more details.
+		// 	initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
+		// 	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
+		// 		*myRayService.Spec.RayClusterSpec.WorkerGroupSpecs[0].Replicas++
+		// 		return k8sClient.Update(ctx, myRayService)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
 
-			// Confirm not switch to a new RayCluster
-			Consistently(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*5, time.Millisecond*500).Should(Equal(initialClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
-			Eventually(
-				getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Status.ActiveServiceStatus.RayClusterName, Namespace: "default"}, myRayCluster),
-				time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayCluster  = %v", myRayCluster.Name)
-		})
+		// 	// Confirm not switch to a new RayCluster
+		// 	Consistently(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*5, time.Millisecond*500).Should(Equal(initialClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// 	Eventually(
+		// 		getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Status.ActiveServiceStatus.RayClusterName, Namespace: "default"}, myRayCluster),
+		// 		time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayCluster  = %v", myRayCluster.Name)
+		// })
 
-		It("should perform a zero-downtime update after a code change.", func() {
-			initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
+		// It("should perform a zero-downtime update after a code change.", func() {
+		// 	initialClusterName, _ := getRayClusterNameFunc(ctx, myRayService)()
 
-			// The cluster shouldn't switch until deployments are finished updating
-			fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.UPDATING, rayv1.ApplicationStatusEnum.DEPLOYING))
+		// 	// The cluster shouldn't switch until deployments are finished updating
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.UPDATING, rayv1.ApplicationStatusEnum.DEPLOYING))
 
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
-				myRayService.Spec.RayClusterSpec.HeadGroupSpec.Template.Spec.Containers[0].Env[1].Value = "UPDATED_VALUE"
-				myRayService.Spec.RayClusterSpec.WorkerGroupSpecs[0].Template.Spec.Containers[0].Env[1].Value = "UPDATED_VALUE"
-				return k8sClient.Update(ctx, myRayService)
-			})
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
+		// 	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
+		// 		myRayService.Spec.RayClusterSpec.HeadGroupSpec.Template.Spec.Containers[0].Env[1].Value = "UPDATED_VALUE"
+		// 		myRayService.Spec.RayClusterSpec.WorkerGroupSpecs[0].Template.Spec.Containers[0].Env[1].Value = "UPDATED_VALUE"
+		// 		return k8sClient.Update(ctx, myRayService)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayService resource")
 
-			Eventually(
-				getPreparingRayClusterNameFunc(ctx, myRayService),
-				time.Second*60, time.Millisecond*500).Should(Not(BeEmpty()), "My new RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
+		// 	Eventually(
+		// 		getPreparingRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*60, time.Millisecond*500).Should(Not(BeEmpty()), "My new RayCluster name  = %v", myRayService.Status.PendingServiceStatus.RayClusterName)
 
-			pendingRayClusterName := myRayService.Status.PendingServiceStatus.RayClusterName
+		// 	pendingRayClusterName := myRayService.Status.PendingServiceStatus.RayClusterName
 
-			Consistently(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*5, time.Millisecond*500).Should(Equal(initialClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// 	Consistently(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*5, time.Millisecond*500).Should(Equal(initialClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
 
-			// The cluster should switch once the deployments are finished updating
-			fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
-			updateHeadPodToRunningAndReady(ctx, pendingRayClusterName)
+		// 	// The cluster should switch once the deployments are finished updating
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING))
+		// 	updateHeadPodToRunningAndReady(ctx, pendingRayClusterName)
 
-			Eventually(
-				getRayClusterNameFunc(ctx, myRayService),
-				time.Second*60, time.Millisecond*500).Should(Equal(pendingRayClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
-		})
+		// 	Eventually(
+		// 		getRayClusterNameFunc(ctx, myRayService),
+		// 		time.Second*60, time.Millisecond*500).Should(Equal(pendingRayClusterName), "My current RayCluster name  = %v", myRayService.Status.ActiveServiceStatus.RayClusterName)
+		// })
 
-		It("should reconcile status correctly when multi-app config type is used.", func() {
-			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
-				Eventually(
-					getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
-					time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
+		// It("should reconcile status correctly when multi-app config type is used.", func() {
+		// 	err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		// 		Eventually(
+		// 			getResourceFunc(ctx, client.ObjectKey{Name: myRayService.Name, Namespace: "default"}, myRayService),
+		// 			time.Second*3, time.Millisecond*500).Should(BeNil(), "My myRayService  = %v", myRayService.Name)
 
-				myRayService.Spec.ServeDeploymentGraphSpec = rayv1.ServeDeploymentGraphSpec{}
-				myRayService.Spec.ServeConfigV2 = testServeConfigV2
-				return k8sClient.Update(ctx, myRayService)
-			})
-			Expect(err).NotTo(HaveOccurred(), "failed to update test RayService serve config")
+		// 		myRayService.Spec.ServeDeploymentGraphSpec = rayv1.ServeDeploymentGraphSpec{}
+		// 		myRayService.Spec.ServeConfigV2 = testServeConfigV2
+		// 		return k8sClient.Update(ctx, myRayService)
+		// 	})
+		// 	Expect(err).NotTo(HaveOccurred(), "failed to update test RayService serve config")
 
-			// Set multi-application status to healthy.
-			healthyStatus := generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING)
-			fakeRayDashboardClient.SetMultiApplicationStatuses(map[string]*utils.ServeApplicationStatus{testServeAppName: &healthyStatus})
-			// Set single-application status to unhealthy.
-			unhealthyStatus := generateServeStatus(rayv1.DeploymentStatusEnum.UNHEALTHY, rayv1.ApplicationStatusEnum.UNHEALTHY)
-			fakeRayDashboardClient.SetSingleApplicationStatus(unhealthyStatus)
+		// 	// Set multi-application status to healthy.
+		// 	healthyStatus := generateServeStatus(rayv1.DeploymentStatusEnum.HEALTHY, rayv1.ApplicationStatusEnum.RUNNING)
+		// 	fakeRayDashboardClient.SetMultiApplicationStatuses(map[string]*utils.ServeApplicationStatus{testServeAppName: &healthyStatus})
+		// 	// Set single-application status to unhealthy.
+		// 	unhealthyStatus := generateServeStatus(rayv1.DeploymentStatusEnum.UNHEALTHY, rayv1.ApplicationStatusEnum.UNHEALTHY)
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(unhealthyStatus)
 
-			// The status should remain healthy, because we have set serveConfigTypeForTesting to MULTI_APP
-			// Then the rayservice controller should pull the deployment statuses from the multi-app API and not the single-app API
-			// So regardless of what the single-app API is returning, the status of the rayservice should remain healthy
-			// because the multi-app API is returning healthy statuses.
-			Consistently(
-				checkServiceHealth(ctx, myRayService),
-				time.Second*5, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
+		// 	// The status should remain healthy, because we have set serveConfigTypeForTesting to MULTI_APP
+		// 	// Then the rayservice controller should pull the deployment statuses from the multi-app API and not the single-app API
+		// 	// So regardless of what the single-app API is returning, the status of the rayservice should remain healthy
+		// 	// because the multi-app API is returning healthy statuses.
+		// 	Consistently(
+		// 		checkServiceHealth(ctx, myRayService),
+		// 		time.Second*5, time.Millisecond*500).Should(BeTrue(), "myRayService status = %v", myRayService.Status)
 
-			// reset states to healthy for following test cases
-			fakeRayDashboardClient.SetSingleApplicationStatus(healthyStatus)
-		})
+		// 	// reset states to healthy for following test cases
+		// 	fakeRayDashboardClient.SetSingleApplicationStatus(healthyStatus)
+		// })
 	})
 })
